requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.0
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.4
django-allauth==0.36.0
django-celery-beat==1.1.1
django-ckeditor==5.5.0
django-js-asset==1.1.0
idna==2.7
kombu==4.2.1
oauthlib==2.1.0
Pillow==5.1.0
python3-openid==3.1.0
pytz==2018.4
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
stripe==2.4.0
urllib3==1.23
vine==1.1.4
